% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Unzip database.csv (if needed), read, and clean data.}
\usage{
clean()
}
\value{
returns a clean tibble ready for anlysis
}
\description{
Unzip database.csv (if needed), read, and clean data.
}
